+++ app/code/community/Bolt/Boltpay/Block/Checkout/Boltpay.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * Magento
+ * Bolt magento plugin
  *
  * NOTICE OF LICENSE
  *
@@ -8,19 +8,10 @@
  * that is bundled with this package in the file LICENSE.txt.
  * It is also available through the world-wide-web at this URL:
  * http://opensource.org/licenses/osl-3.0.php
- * If you did not receive a copy of the license and are unable to
- * obtain it through the world-wide-web, please send an email
- * to license@magento.com so we can send you a copy immediately.
- *
- * DISCLAIMER
- *
- * Do not edit or add to this file if you wish to upgrade the Bolt extension
- * to a newer versions in the future. If you wish to customize this extension
- * for your needs please refer to http://www.magento.com for more information.
  *
  * @category   Bolt
  * @package    Bolt_Boltpay
- * @copyright  Copyright (c) 2018 Bolt Financial, Inc (http://www.bolt.com)
+ * @copyright  Copyright (c) 2018 Bolt Financial, Inc (https://www.bolt.com)
  * @license    http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
 
@@ -57,6 +48,12 @@ class Bolt_Boltpay_Block_Checkout_Boltpay extends Mage_Checkout_Block_Onepage_Re
     const CSS_SUFFIX = 'bolt-css-suffix';
 
     /**
+     * Javascript compatible string to signify the cart needs a shipping address before
+     * Bolt Cart JS can be generated.
+     */
+    const NEEDS_SHIPPING_JS = '"needs_shipping";';
+
+    /**
      * Set the connect javascript url to production or sandbox based on store config settings
      */
     public function _construct()
@@ -86,11 +83,11 @@ class Bolt_Boltpay_Block_Checkout_Boltpay extends Mage_Checkout_Block_Onepage_Re
      */
     private function createBoltOrder($quote, $multipage)
     {
-
         // Load the required helper class
         $boltHelper = Mage::helper('boltpay/api');
 
-        $items = $this->getItems();
+        $items = $quote->getAllVisibleItems();
+
         if (empty($items)) return json_decode('{"token" : ""}');
 
         // Generates order data for sending to Bolt create order API.
@@ -107,18 +104,18 @@ class Bolt_Boltpay_Block_Checkout_Boltpay extends Mage_Checkout_Block_Onepage_Re
      * In BoltCheckout.configure success callback the order is saved in additional ajax call to
      * Bolt_Boltpay_OrderController save action.
      *
-     * @param bool $multipage       Is checkout type Multi-Page Checkout, the default is true, set to false for One Page Checkout
+     * @param string $checkoutType  'multi-page' | 'one-page' | 'admin'
      * @return string               BoltCheckout javascript
      */
-    public function getCartDataJs($multipage = true)
+    public function getCartDataJs($checkoutType = 'multi-page')
     {
         try {
             // Get customer and cart session objects
             $customerSession = Mage::getSingleton('customer/session');
-            $session = Mage::getSingleton('checkout/session');
+            $session = ($checkoutType === 'admin') ? Mage::getSingleton('adminhtml/session_quote') : Mage::getSingleton('checkout/session');
 
             /* @var Mage_Sales_Model_Quote $sessionQuote */
-            $sessionQuote = $session->getQuote();
+            $sessionQuote =  $session->getQuote();
 
             /* @var Bolt_Boltpay_Helper_Api $boltHelper */
             $boltHelper = Mage::helper('boltpay/api');
@@ -126,7 +123,7 @@ class Bolt_Boltpay_Block_Checkout_Boltpay extends Mage_Checkout_Block_Onepage_Re
             ///////////////////////////////////////////////////////////////
             // Populate hints data from quote or customer shipping address.
             //////////////////////////////////////////////////////////////
-            $hintData = $this->getAddressHints($customerSession, $sessionQuote);
+            $hintData = $this->getAddressHints($customerSession, $sessionQuote, $checkoutType);
             ///////////////////////////////////////////////////////////////
 
             // Check if cart contains at least one item.
@@ -154,7 +151,7 @@ class Bolt_Boltpay_Block_Checkout_Boltpay extends Mage_Checkout_Block_Onepage_Re
                 }
                 ///////////////////////////////////////////////////////////////////////////////////////
 
-                if($multipage) {
+                if($checkoutType === 'multi-page') {
                     // Resets shipping rate
                     $shippingMethod = $sessionQuote->getShippingAddress()->getShippingMethod();
                     $boltHelper->applyShippingRate($sessionQuote, null);
@@ -163,6 +160,8 @@ class Bolt_Boltpay_Block_Checkout_Boltpay extends Mage_Checkout_Block_Onepage_Re
                 // Call Bolt create order API
                 $orderCreationResponse = json_decode('{"token" : ""}');
                 try {
+                    $isMultiPage = $checkoutType === 'multi-page';
+
                     /////////////////////////////////////////////////////////////////////////////////
                     // We create a copy of the quote that is immutable by the customer/frontend
                     // Bolt saves this quote to the database at Magento-side order save time.
@@ -170,7 +169,7 @@ class Bolt_Boltpay_Block_Checkout_Boltpay extends Mage_Checkout_Block_Onepage_Re
                     // Only shipping, tax and discounts can change, and only if the shipping, tax
                     // and discount calculations change on the Magento server
                     ////////////////////////////////////////////////////////////////////////////////
-                    $immutableQuote = $boltHelper->cloneQuote($sessionQuote, $multipage);
+                    $immutableQuote = $boltHelper->cloneQuote($sessionQuote, $isMultiPage);
 
                     ///////////////////////////////////////////////////////////////////////////////
                     // if we are missing a shipping method in one-page for a non-virtual cart
@@ -178,17 +177,16 @@ class Bolt_Boltpay_Block_Checkout_Boltpay extends Mage_Checkout_Block_Onepage_Re
                     // BoltCheckout.configure will be made because a shipping address is required
                     // under these conditions in order to make that call
                     ///////////////////////////////////////////////////////////////////////////////
-                    if (!$multipage && !$immutableQuote->isVirtual() && !$immutableQuote->getShippingAddress()->getShippingMethod() ) {
-                        return '"needs_shipping";';
+                    if (!$isMultiPage && !$immutableQuote->isVirtual() && !$immutableQuote->getShippingAddress()->getShippingMethod() ) {
+                        return self::NEEDS_SHIPPING_JS;
                     }
 
-                    $orderCreationResponse = $this->createBoltOrder($immutableQuote, $multipage);
-
+                    $orderCreationResponse = $this->createBoltOrder($immutableQuote, $isMultiPage);
                 } catch (Exception $e) {
                     Mage::helper('boltpay/bugsnag')->notifyException(new Exception($e));
                 }
 
-                if ($multipage) {
+                if ($checkoutType === 'multi-page') {
                     $boltHelper->applyShippingRate($sessionQuote, $shippingMethod);
                 }
 
@@ -217,24 +215,106 @@ class Bolt_Boltpay_Block_Checkout_Boltpay extends Mage_Checkout_Block_Onepage_Re
 
             //////////////////////////////////////////////////////////////////////////
             // Format the success and save order urls for the javascript returned below.
-            $successUrl    = $this->getUrl(Mage::getStoreConfig('payment/boltpay/successpage'));
+            $successUrl   = $this->getUrl(Mage::getStoreConfig('payment/boltpay/successpage'));
             $saveOrderUrl = $this->getUrl('boltpay/order/save');
 
             //////////////////////////////////////////////////////
             // Collect the event Javascripts
+            // We execute these events as early as possible, typically
+            // before Bolt defined event JS to give merchants the
+            // opportunity to do full overrides
             //////////////////////////////////////////////////////
-            $check = Mage::getStoreConfig('payment/boltpay/check');
-            $onCheckoutStart = Mage::getStoreConfig('payment/boltpay/on_checkout_start');
-            $onShippingDetailsComplete = Mage::getStoreConfig('payment/boltpay/on_shipping_details_complete');
-            $onShippingOptionsComplete = Mage::getStoreConfig('payment/boltpay/on_shipping_options_complete');
-            $onPaymentSubmit = Mage::getStoreConfig('payment/boltpay/on_payment_submit');
-            $success = Mage::getStoreConfig('payment/boltpay/success');
-            $close = Mage::getStoreConfig('payment/boltpay/close');
+            if (($checkoutType === 'admin') && !Mage::getStoreConfig('payment/boltpay/use_javascript_in_admin')) {
+                $check = '';
+                $onCheckoutStart = '';
+                $onShippingDetailsComplete = '';
+                $onShippingOptionsComplete = '';
+                $onPaymentSubmit = '';
+                $success = '';
+                $close = '';
+            } else {
+                $check = Mage::getStoreConfig('payment/boltpay/check');
+                $onCheckoutStart = Mage::getStoreConfig('payment/boltpay/on_checkout_start');
+                $onShippingDetailsComplete = Mage::getStoreConfig('payment/boltpay/on_shipping_details_complete');
+                $onShippingOptionsComplete = Mage::getStoreConfig('payment/boltpay/on_shipping_options_complete');
+                $onPaymentSubmit = Mage::getStoreConfig('payment/boltpay/on_payment_submit');
+                $success = Mage::getStoreConfig('payment/boltpay/success');
+                $close = Mage::getStoreConfig('payment/boltpay/close');
+            }
 
             //////////////////////////////////////////////////////
             // Generate and return BoltCheckout javascript.
             //////////////////////////////////////////////////////
             $immutableQuoteId = $immutableQuote->getId();
+
+            $checkForAdmin = '';
+            $onSuccessCallback = '';
+            $onClose = '';
+            if ($checkoutType === 'admin') {
+                $checkForAdmin = "if ((typeof editForm !== 'undefined') && (typeof editForm.validate === 'function')) {
+                    var bolt_hidden = document.getElementById('boltpay_payment_button');
+                    bolt_hidden.classList.remove('required-entry');
+
+                    var is_valid = true;
+
+                    if (!editForm.validate()) {
+                        is_valid = false;
+                    } else {
+                        var shipping_method = $$('input:checked[type=\"radio\"][name=\"order[shipping_method]\"]')[0];
+                        if (typeof shipping_method === 'undefined') {
+                            alert('Please select a shipping method.');
+                            is_valid = false;
+                        }
+                    }
+
+                    bolt_hidden.classList.add('required-entry');
+                    return is_valid;
+                }";
+                $onSuccessCallback = "function(transaction, callback) {
+                    $success
+                    
+                    var input = document.createElement('input');
+                    input.setAttribute('type', 'hidden');
+                    input.setAttribute('name', 'bolt_reference');
+                    input.setAttribute('value', transaction.reference);
+                    document.getElementById('edit_form').appendChild(input);
+                    
+                    // order and order.submit should exist for admin            
+                    if ((typeof order !== 'undefined' ) && (typeof order.submit === 'function')) {
+                        order_completed = true;
+                        callback();
+                    }
+                }";
+                $onClose = "if (order_completed && (typeof order !== 'undefined' ) && (typeof order.submit === 'function')) {
+                    $close
+                    var bolt_hidden = document.getElementById('boltpay_payment_button');
+                    bolt_hidden.classList.remove('required-entry');
+                    order.submit();
+                 }";
+            } else {
+                $onSuccessCallback = "function(transaction, callback) {
+                    new Ajax.Request(
+                        '$saveOrderUrl',
+                        {
+                            method:'post',
+                            onSuccess:
+                                function() {
+                                    $success
+                                    order_completed = true;
+                                    callback();
+                                },
+                            parameters: 'reference='+transaction.reference
+                        }
+                    );
+                }";
+                $onClose = "if (typeof bolt_checkout_close === 'function') {
+                    // used internally to set overlay in firecheckout
+                    bolt_checkout_close();
+                 }
+                 if (order_completed) {
+                    location.href = '$successUrl';
+                 }";
+            }
 
             return ("
                 var json_cart = $jsonCart;
@@ -247,15 +327,16 @@ class Bolt_Boltpay_Block_Checkout_Boltpay extends Mage_Checkout_Block_Onepage_Re
                     $jsonHints,
                     {
                       check: function() {
+                        $check
+                        $checkForAdmin
                         if (isEmptyQuote) {
-                            alert('{$boltHelper->__('Please add some product to cart')}');
+                            alert('{$boltHelper->__('Your shopping cart is empty. Please add products to the cart.')}');
                             return false;
                         }
                         if (!json_cart.orderToken) {
                             alert(json_cart.error);
                             return false;
                         }
-                        $check
                         return true;
                       },
                       
@@ -281,31 +362,11 @@ class Bolt_Boltpay_Block_Checkout_Boltpay extends Mage_Checkout_Block_Onepage_Re
                         $onPaymentSubmit
                       },
                       
-                      success: function(transaction, callback) {
-                        new Ajax.Request(
-                            '$saveOrderUrl',
-                            {
-                                method:'post',
-                                onSuccess: 
-                                    function() {
-                                        $success
-                                        order_completed = true;
-                                        callback();  
-                                    },
-                                parameters: 'reference='+transaction.reference
-                            }
-                        );
-                      },
-                      
+                      success: $onSuccessCallback,
+
                       close: function() {
                          $close
-                         if (typeof bolt_checkout_close === 'function') {
-                            // used internally to set overlay in firecheckout
-                            bolt_checkout_close();
-                         }
-                         if (order_completed) {   
-                            location.href = '$successUrl';
-                         }
+                         $onClose
                       }
                     }
                 );"
@@ -320,40 +381,45 @@ class Bolt_Boltpay_Block_Checkout_Boltpay extends Mage_Checkout_Block_Onepage_Re
      * Get address data for sending as hints.
      *
      * @param $session      Customer session
+     * @param string $checkoutType  'multi-page' | 'one-page' | 'admin'
+     *
      * @return array        hints data
      */
-    private function getAddressHints($session, $quote)
+    private function getAddressHints($session, $quote, $checkoutType)
     {
 
         $hints = array();
 
-        ///////////////////////////////////////////////////////////////
+        /////////////////////////////////////////////////////////////////////////
         // Check if the quote shipping address is set,
         // otherwise use customer shipping address for logged in users.
-        ///////////////////////////////////////////////////////////////
+        /////////////////////////////////////////////////////////////////////////
         $address = $quote->getShippingAddress();
-
-        if ($session && $session->isLoggedIn()) {
-            /** @var Mage_Customer_Model_Customer $customer */
-            $customer = Mage::getModel('customer/customer')->load($session->getId());
-            $address = $customer->getPrimaryShippingAddress();
-            $email = $customer->getEmail();
+        if (!$address->getStreet1()) {
+            if ( $session && $session->isLoggedIn()) {
+                /** @var Mage_Customer_Model_Customer $customer */
+                $customer = Mage::getModel('customer/customer')->load($session->getId());
+                $address = $customer->getPrimaryShippingAddress();
+                $hints['email'] = $customer->getEmail();
+            }
         }
 
-        /////////////////////////////////////////////////////////////////////////
-        // If address exists populate the hints array with existing address data.
-        /////////////////////////////////////////////////////////////////////////
-        if ($address) {
-            if (@$email)                   $hints['email']        = $email;
-            if (@$address->getFirstname()) $hints['firstName']    = $address->getFirstname();
-            if (@$address->getLastname())  $hints['lastName']     = $address->getLastname();
-            if (@$address->getStreet1())   $hints['addressLine1'] = $address->getStreet1();
-            if (@$address->getStreet2())   $hints['addressLine2'] = $address->getStreet2();
-            if (@$address->getCity())      $hints['city']         = $address->getCity();
-            if (@$address->getRegion())    $hints['state']        = $address->getRegion();
-            if (@$address->getPostcode())  $hints['zip']          = $address->getPostcode();
-            if (@$address->getTelephone()) $hints['phone']        = $address->getTelephone();
-            if (@$address->getCountryId()) $hints['country']      = $address->getCountryId();
+        // If address value exists populate the hints array with existing address data.
+        if ($address->getEmail())     $hints['email']        = $address->getEmail();
+        if ($address->getFirstname()) $hints['firstName']    = $address->getFirstname();
+        if ($address->getLastname())  $hints['lastName']     = $address->getLastname();
+        if ($address->getStreet1())   $hints['addressLine1'] = $address->getStreet1();
+        if ($address->getStreet2())   $hints['addressLine2'] = $address->getStreet2();
+        if ($address->getCity())      $hints['city']         = $address->getCity();
+        if ($address->getRegion())    $hints['state']        = $address->getRegion();
+        if ($address->getPostcode())  $hints['zip']          = $address->getPostcode();
+        if ($address->getTelephone()) $hints['phone']        = $address->getTelephone();
+        if ($address->getCountryId()) $hints['country']      = $address->getCountryId();
+
+
+        if ($checkoutType === 'admin') {
+            $hints['email'] = Mage::getSingleton('admin/session')->getOrderShippingAddress()['email'];
+            $hints['virtual_terminal_mode'] = true;
         }
 
         return array( "prefill" => $hints );
@@ -533,17 +599,28 @@ class Bolt_Boltpay_Block_Checkout_Boltpay extends Mage_Checkout_Block_Onepage_Re
     /**
      * Returns the One Page / Multi-Page checkout Publishable key.
      *
-     * @param bool $multipage
-     * @return string
+     * @param  string $checkoutType  'multi-page' | 'one-page' | 'admin'
+     *
+     * @return string the publishable key associated with the type of checkout requested
      */
-    function getPublishableKey($multipage = true)
+    function getPublishableKey($checkoutType = 'multi-page')
     {
         /** @var Bolt_Boltpay_Helper_Data $hlp */
         $hlp = $this->helper('boltpay');
 
-        return $multipage
-            ? $hlp->getPublishableKeyMultiPageKey(true)
-            : $hlp->getPublishableKeyOnePageKey(true);
+        switch ($checkoutType) {
+            case 'multi-page':
+            case 'multipage':
+                return $hlp->getPublishableKeyMultiPage();
+            case 'back-office':
+            case 'backoffice':
+            case 'admin':
+                return $hlp->getPublishableKeyBackOffice();
+            case 'one-page':
+            case 'onepage':
+            default:
+                return $hlp->getPublishableKeyOnePage();
+        }
     }
 
     /**
@@ -643,7 +720,7 @@ class Bolt_Boltpay_Block_Checkout_Boltpay extends Mage_Checkout_Block_Onepage_Re
         $routeName = $this->getRequest()->getRouteName();
         $controllerName = $this->getRequest()->getControllerName();
 
-        $isAllowed = ($routeName === 'checkout' && $controllerName === 'cart');
+        $isAllowed = ($routeName === 'checkout' && $controllerName === 'cart') || $routeName == 'firecheckout';
 
         return $isAllowed;
     }