+++ app/code/community/Bolt/Boltpay/controllers/ShippingController.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * Magento
+ * Bolt magento plugin
  *
  * NOTICE OF LICENSE
  *
@@ -8,19 +8,10 @@
  * that is bundled with this package in the file LICENSE.txt.
  * It is also available through the world-wide-web at this URL:
  * http://opensource.org/licenses/osl-3.0.php
- * If you did not receive a copy of the license and are unable to
- * obtain it through the world-wide-web, please send an email
- * to license@magento.com so we can send you a copy immediately.
- *
- * DISCLAIMER
- *
- * Do not edit or add to this file if you wish to upgrade the Bolt extension
- * to a newer versions in the future. If you wish to customize this extension
- * for your needs please refer to http://www.magento.com for more information.
  *
  * @category   Bolt
  * @package    Bolt_Boltpay
- * @copyright  Copyright (c) 2018 Bolt Financial, Inc (http://www.bolt.com)
+ * @copyright  Copyright (c) 2018 Bolt Financial, Inc (https://www.bolt.com)
  * @license    http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
 
@@ -61,6 +52,12 @@ class Bolt_Boltpay_ShippingController extends Mage_Core_Controller_Front_Action
 
             $shippingAddress = $requestData->shipping_address;
 
+            if (!$this->isPOBoxAllowed() && $this->doesAddressContainPOBox($shippingAddress->street_address1, $shippingAddress->street_address2)) {
+                $errorDetails = array('code' => 6101, 'message' => Mage::helper('boltpay')->__('Address with P.O. Box is not allowed.'));
+                return $this->getResponse()->setHttpResponseCode(403)
+                    ->setBody(json_encode(array('status' => 'failure','error' => $errorDetails)));
+            }
+
             $region = Mage::getModel('directory/region')->loadByName($shippingAddress->region, $shippingAddress->country_code)->getCode();
 
             $addressData = array(
@@ -77,9 +74,9 @@ class Bolt_Boltpay_ShippingController extends Mage_Core_Controller_Front_Action
             );
 
             $quoteId = $requestData->cart->order_reference;
+
             /* @var Mage_Sales_Model_Quote $quote */
             $quote = Mage::getModel('sales/quote')->loadByIdWithoutStore($quoteId);
-
 
             /***********************/
             // Set session quote to real customer quote
@@ -177,7 +174,12 @@ class Bolt_Boltpay_ShippingController extends Mage_Core_Controller_Front_Action
 
             $this->getResponse()->setBody($response);
         } catch (Exception $e) {
-            Mage::helper('boltpay/bugsnag')->notifyException($e);
+            $metaData = array();
+            if (isset($quote)){
+                $metaData['quote'] = var_export($quote->debug(), true);
+            }
+
+            Mage::helper('boltpay/bugsnag')->notifyException($e, $metaData);
             throw $e;
         }
     }
@@ -205,13 +207,13 @@ class Bolt_Boltpay_ShippingController extends Mage_Core_Controller_Front_Action
             $geoLocationAddress = $this->cleanEmptyAddressField($geoLocationAddress);
 
             // ----------^_^----------- //
-            $shippingAddress = [
+            $shippingAddress = array(
                 'city'       => @$shippingAddressOriginal['city'],
                 'region'     => @$shippingAddressOriginal['region'],
                 'region_id'  => @$shippingAddressOriginal['region_id'] ? $shippingAddressOriginal['region_id'] : null,
                 'postcode'   => @$shippingAddressOriginal['postcode'],
                 'country_id' => @$shippingAddressOriginal['country_id'],
-            ];
+            );
             unset($shippingAddressOriginal);
 
             $addressData = $this->mergeAddressData($geoLocationAddress, $shippingAddress);
@@ -382,5 +384,29 @@ class Bolt_Boltpay_ShippingController extends Mage_Core_Controller_Front_Action
         }
 
         return md5($cacheIdentifier);
+    }
+
+    /**
+     * @return mixed
+     */
+    protected function isPOBoxAllowed()
+    {
+        return Mage::getStoreConfig('payment/boltpay/allow_po_box');
+    }
+
+    /**
+     * @param $address1
+     * @param $address2
+     * @return mixed
+     */
+    public function doesAddressContainPOBox($address1, $address2 = null)
+    {
+        $poBoxRegex = '/^\s*((P(OST)?.?\s*(O(FF(ICE)?)?|B(IN|OX))+.?\s+(B(IN|OX))?)|B(IN|OX))/i';
+
+        if (preg_match($poBoxRegex, $address1) || preg_match($poBoxRegex, $address2)) {
+            return true;
+        }
+
+        return false;
     }
 }